requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
slowapi==0.1.9
insightface==0.7.3
numpy==1.23.5
torch==2.0.1
PyYAML==6.0
opencv-python==4.8.0.76
onnx==1.17.0
onnxruntime==1.19.0
aiofiles==23.2.1